// app/components/chat/style.module.css.d.ts
declare const styles: {
  answerIcon: string;
  typingIcon: string;
  questionIcon: string;
  answer: string;
  question: string;
  answerWrap: string;
  itemOperation: string;
  textArea: string;
  count: string;
  sendBtn: string;
  answerContainer: string;
  
  // 代码高亮样式
  codeBlock: string;
  codeContent: string;
  keyword: string;
  table: string;
  column: string;
  function: string;
  number: string;
  string: string;
  operator: string;
};

export default styles; 